+++ VcRedist/Public/Get-VcList.ps1
@@ -13,7 +13,7 @@ Function Get-VcList {
             Twitter: @stealthpuppy
 
         .LINK
-            https://docs.stealthpuppy.com/docs/vcredist/usage/getting-the-vcredist-list
+            https://stealthpuppy.com/VcRedist/get-vclist.html
 
         .PARAMETER Manifest
             The JSON file that contains the details about the Visual C++ Redistributables. This must be in the expected format.
@@ -65,7 +65,7 @@ Function Get-VcList {
             Returns a list of the Visual C++ Redistributables listed in the external manifest VisualCRedistributables.json.
     #>
     [OutputType([System.Management.Automation.PSObject])]
-    [CmdletBinding(DefaultParameterSetName = 'Manifest', HelpURI = "https://docs.stealthpuppy.com/docs/vcredist/usage/getting-the-vcredist-list")]
+    [CmdletBinding(DefaultParameterSetName = 'Manifest', HelpURI = "https://stealthpuppy.com/VcRedist/get-vclist.html")]
     Param (
         [Parameter(Mandatory = $False, Position = 0, ValueFromPipeline, ParameterSetName = 'Manifest')]
         [ValidateNotNull()]