+++ VcRedist/Public/Install-VcRedist.ps1
@@ -11,7 +11,7 @@ Function Install-VcRedist {
             Twitter: @stealthpuppy
 
         .LINK
-            https://docs.stealthpuppy.com/docs/vcredist/usage/installing-the-redistributables
+            https://stealthpuppy.com/VcRedist/install-vcredist.html
 
         .PARAMETER VcList
             An array containing details of the Visual C++ Redistributables from Get-VcList.
@@ -43,7 +43,7 @@ Function Install-VcRedist {
             Description:
             Installs all supported Visual C++ Redistributables using a completely silent install.
     #>
-    [CmdletBinding(SupportsShouldProcess = $True, HelpURI = "https://docs.stealthpuppy.com/docs/vcredist/usage/installing-the-redistributables")]
+    [CmdletBinding(SupportsShouldProcess = $True, HelpURI = "https://stealthpuppy.com/VcRedist/install-vcredist.html")]
     [OutputType([System.Management.Automation.PSObject])]
     Param (
         [Parameter(Mandatory = $True, Position = 0, ValueFromPipeline)]