+++ VcRedist/Public/Save-VcRedist.ps1
@@ -23,7 +23,7 @@ Function Save-VcRedist {
             Twitter: @stealthpuppy
 
         .LINK
-            https://docs.stealthpuppy.com/docs/vcredist/usage/downloading-the-redistributables
+            https://stealthpuppy.com/VcRedist/save-vcredist.html
 
         .PARAMETER VcList
             Specifies the array that lists the Visual C++ Redistributables to download
@@ -72,7 +72,7 @@ Function Save-VcRedist {
             Downloads the 2010, 2012, 2013, and 2019 Visual C++ Redistributables to C:\Redist using the proxy server 'proxy.domain.local'
         #>
     [Alias("Get-VcRedist")]
-    [CmdletBinding(SupportsShouldProcess = $True, HelpURI = "https://docs.stealthpuppy.com/docs/vcredist/usage/downloading-the-redistributables")]
+    [CmdletBinding(SupportsShouldProcess = $True, HelpURI = "https://stealthpuppy.com/VcRedist/save-vcredist.html")]
     [OutputType([System.Management.Automation.PSObject])]
     Param (
         [Parameter(Mandatory = $True, Position = 0, ValueFromPipeline)]